requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
anthropic>=0.30.0
py-clob-client>=0.16.0
flask>=3.0.0
psycopg2-binary>=2.9.9
